adding a new addon "get"
