+++ system/api/server.go
@@ -1,11 +1,6 @@
 package api
 
-import (
-	"bytes"
-	"fmt"
-	"log"
-	"net/http"
-)
+import "net/http"
 
 // Run adds Handlers to default http listener for API
 func Run() {
@@ -16,24 +11,4 @@ func Run() {
 	http.HandleFunc("/api/content", CORS(Record(contentHandler)))
 
 	http.HandleFunc("/api/content/external", CORS(Record(externalContentHandler)))
-}
-
-// ContentAll retrives all items from the HTTP API within the provided namespace
-func ContentAll(namespace string) []byte {
-	endpoint := "http://0.0.0.0:8080/api/contents?type="
-	buf := []byte{}
-	r := bytes.NewReader(buf)
-	req, err := http.NewRequest(http.MethodGet, endpoint+namespace, r)
-	if err != nil {
-		log.Println("Error creating request for reference from:", namespace)
-		return nil
-	}
-
-	c := http.Client{}
-	res, err := c.Do(req)
-	defer res.Body.Close()
-
-	fmt.Println(res, string(buf))
-
-	return buf
 }
